--- app/src/main/java/com/example/kotlin_cardgame/adapter/FragmentStatePagerAdapter.kt
@@ -1,32 +1,0 @@
-package com.example.kotlin_cardgame.adapter
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.kotlin_cardgame.fragment.GameFlagment
-import com.example.kotlin_cardgame.fragment.SettingFlagment
-
-class FragmentStatePagerAdapter(fragmentActivity: FragmentActivity) : FragmentStateAdapter(fragmentActivity) {
-    var fragments : ArrayList<Fragment> = ArrayList()
-
-    override fun getItemCount(): Int {
-        return fragments.size
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return when (position) {
-            0 -> GameFlagment()
-            else -> SettingFlagment()
-        }
-    }
-
-    fun addFragment(fragment: Fragment) {
-        fragments.add(fragment)
-        notifyItemInserted(fragments.size - 1)
-    }
-
-    fun removeFragment() {
-        fragments.removeLast()
-        notifyItemRemoved(fragments.size)
-    }
-}
