+++ app/src/main/java/com/example/kotlin_cardgame/fragment/GameFragment.kt
@@ -1,0 +1,19 @@
+package com.example.kotlin_cardgame.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.kotlin_cardgame.R
+
+
+class GameFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_game, container, false)
+    }
+}
