+++ app/src/main/res/values/strings.xml
@@ -6,5 +6,6 @@
     <string name="imagebutton4">imageButton4</string>
     <string name="imageview">imageView</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="hello_game_fragment">Game fragment</string>
+    <string name="hello_setting_fragment">Setting fragment</string>
 </resources>
